Fix: ToB-21 (unsafe casts) (#1457)
* Add ChainContext library

* Add coverage for the new library

* Remove unsafe casts for getting chain context

* Safe cast for attestation nonces

* Make `stateIndex` uint8

* Adjust tests with uint8 stateIndex

* Add safe casts for 32 bits indexes

* Document dispute index unsafe cast

* Document unsafe casts in `memory` libs

* Document unsafe casts which are actually upcasts

* Document remaining unsafe casts

* Chore: better formatting

* Add SafeCast for `encodeTips256()`

* Bound gasDrop/gasLimit max values to avoid overflow in tests

* Revert changes that are already in #1430

* Better docs for snapshot merkle root

* Chore: explicit reverts when bounding in tests

* Add tests with potential context overflows

+++ packages/contracts-core/test/mocks/hubs/ExecutionHubMock.t.sol
@@ -13,7 +13,7 @@ contract ExecutionHubMock is BaseMock, IExecutionHub {
         bytes memory msgPayload,
         bytes32[] calldata originProof,
         bytes32[] calldata snapProof,
-        uint256 stateIndex,
+        uint8 stateIndex,
         uint64 gasLimit
     ) external {}
